Send separate report for layer/mod-tap keypress events
Rolling presses of layer and mod tap keys togehter with a regular key inside the tapping term
produced a single keyboard report for both keypress events. Depending on the keycode this
gave the regular keycode precedence over the tapped key altough it was pressed after the tap keys.
This led to sublte spelling errors with flipped character positions when
typing fast. To fix this we send separate keyboard reports for keypress
events for tapped keys.